Extracted out a bookmarker for general usage
