--- ingestors/ingestor.go
@@ -1,10 +1,0 @@
-package ingestors
-
-import (
-	"github.com/librariesio/depper/data"
-)
-
-type Ingestor interface {
-	Schedule() string
-	Ingest() []data.PackageVersion
-}
