--- ingestors/streaming_ingestor.go
@@ -1,9 +1,0 @@
-package ingestors
-
-import (
-	"github.com/librariesio/depper/data"
-)
-
-type StreamingIngestor interface {
-	Ingest(chan data.PackageVersion)
-}
