--- ingestors/ttler.go
@@ -1,7 +1,0 @@
-package ingestors
-
-import "time"
-
-type TTLer interface {
-	TTL() time.Duration
-}
